+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.0
+### Bug Fixes
+Fix lazyRepeater and lazyModule to automatic trigger inView until all the elements in the viewPort are visible.
+
+### Features
+Add lazyIf parameter on lazyRepeater directive.
+
 # 0.1.1
 ### Bug Fixes
 Fix fickering of the placeholder on lazy-repeater when number of items is smaller than threshold.